+++ python/guiderActor/gimg/guiderImage.py
@@ -34,6 +34,10 @@ class Fiber(object):
         self.radius = r
         self.illrad = illr
 
+        # Flexure values. To be updated, if needed, when we measure the LED position.
+        self.xFlex = 0.0
+        self.yFlex = 0.0
+
         self.reset_star_values()
 
         self.gProbe = None
@@ -57,6 +61,9 @@ class Fiber(object):
         self.dy = np.nan
         self.dRA = np.nan
         self.dDec = np.nan
+
+        self.xFlex = 0.0
+        self.yFlex = 0.0
 
     def __str__(self):
         return ('Fiber id %i: center %g,%g; star %g,%g; radius %g' %
@@ -493,6 +500,8 @@ class GuiderImageAnalysis(object):
                   ('fiberid', None,     'I', None),
                   ('xCenter', 'xcen',   'E', pixunit),
                   ('yCenter', 'ycen',   'E', pixunit),
+                  ('xFlex',   None,     'E', pixunit),
+                  ('yFlex',   None,     'E', pixunit),
                   ('xstar',   'xs',     'E', pixunit),
                   ('ystar',   'ys',     'E', pixunit),
                   ('dx',      None,     'E', 'residual in mm, guidercam frame'),
@@ -773,6 +782,8 @@ class GuiderImageAnalysis(object):
         except Exception as e:
             self.cmd.warn('text=%s'%qstr('PyGuide.starShape failed on fiber %d with Exception: %s.'%(fiber.fiberid,e)))
 
+        return fiber
+
     def _find_stars_gcam(self, image, mask, fibers):
         """Find the stars in a processed gcamera image."""
         ccdInfo = PyGuide.CCDInfo(self.imageBias,self.readNoise,self.ccdGain)
@@ -797,14 +808,13 @@ class GuiderImageAnalysis(object):
 
             stamp = np.s_[y0:y1, x0:x1]
 
-            tritium_offset = []  # Stores flexure measurements using the tritium source(s)
+            tritium_offsets = []  # Stores flexure measurements using the tritium source(s)
 
             # use a medium threshold, since the stars might not be that bright when acquiring
             try:
-                # TODO: once the mask includes the tritium source we can remove this and use
-                # a single statement for all fibres. (JSG)
-                if fiber.gProbe.tritium:
-                    stars = PyGuide.findStars(image[stamp], None, None, ccdInfo, thresh=2)[0]
+                if fiber.gProbe.tritium and fiber.gProbe.disabled is False:
+                    stars = PyGuide.findStars(image[stamp], None, saturated[stamp],
+                                              ccdInfo, thresh=2)[0]
                 else:
                     stars = PyGuide.findStars(image[stamp], good_mask[stamp],
                                               saturated[stamp], ccdInfo, thresh=2)[0]
@@ -815,36 +825,26 @@ class GuiderImageAnalysis(object):
                 # For not trititum sources we call _set_fiber_star to set xstar/ystar in the
                 # fibre to the values measured by PyGuide. For trititum sources, we compare it
                 # with the value in the fibre and determine an offset
-                if not fiber.gProbe.tritium:
+                if not fiber.gProbe.tritium and fiber.gProbe.disabled is False:
                     self._set_fiber_star(fiber, stars, image[stamp], good_mask[stamp],
                                          stampFrameCoords)
                 else:
-                    if len(stars) == 0:
-                        self.cmd.warn('text="PyGuide.findStars did not '
-                                      'find a source for fiber {:d}."'.format(fiber.fiberid))
-                        continue
+                    if self._set_fiber_star(fiber, stars, image[stamp], None, stampFrameCoords):
+                        tritium_offsets.append([fiber.xs - fiber.xcen, fiber.ys - fiber.ycen])
 
-                    star = stars[0]
-
-                    xs = stampFrameCoords[0] + star.xyCtr[0]
-                    ys = stampFrameCoords[1] + star.xyCtr[1]
-
-                    tritium_offset.append([xs - fiber.xcen, ys - fiber.ycen])
-
-                    # TODO: maybe we should output a measurement of the FWHM of
-                    # the LED here as well.
-
-        if len(tritium_offset) > 0:
-            tritium_offset = np.array(tritium_offset)
-            tritium_offset = tritium_offset.mean(axis=0)
+        if len(tritium_offsets) > 0:
+            tritium_offsets = np.array(tritium_offsets)
+            tritium_offset = tritium_offsets.mean(axis=0)
             if np.any(np.abs(tritium_offset) > 0.1):
                 self.cmd.warn('text="tritium sources detected an offset of '
                               '({:.1f}, {:.1f}) pix."'.format(tritium_offset[0],
                                                               tritium_offset[1]))
 
                 for fiber in fibers:
-                    fiber.xcen += tritium_offset[0]
-                    fiber.ycen += tritium_offset[1]
+                    fiber.xFlex = tritium_offset[0]
+                    fiber.yFlex = tritium_offset[1]
+                    fiber.xcen += fiber.xFlex
+                    fiber.ycen += fiber.yFlex
 
         self.fibers = fibers
 
@@ -1068,7 +1068,7 @@ class GuiderImageAnalysis(object):
         self.currentDarkName = darkFileName
     #...
 
-    def _find_fibers_in_flat(self, image, flatFileName, gprobes, hdr):
+    def _find_fibers_in_flat(self, image, flatFileName, gprobes, hdr, force=False):
         """Identify the fibers in a flat image."""
         # TODO: FIXME -- we KNOW the area (ie, the number of pixels) of the
         # fibers -- we could choose the threshold appropriately.
@@ -1243,7 +1243,7 @@ class GuiderImageAnalysis(object):
             self.cmd.diag('text=%s'%qstr('Fiber id %i at (%.1f, %.1f)' % (f.fiberid, f.xcen-dx, f.ycen-dy)))
 
         # Now we deal with tritium/LED sources. We use PyGuide since these sources are point-like.
-        n_tritium = 1
+        n_tritium = 0
         for gprobe in gprobes.values():
 
             if not gprobe.tritium:
@@ -1275,6 +1275,7 @@ class GuiderImageAnalysis(object):
             except Exception as ee:
                 self.cmd.warn('text="failed to find centroid for tritium gprobe {}: {}"'
                               .format(gprobe.id, ee))
+                gprobe.disabled = True
             else:
                 fiber_x = (stampFrameCoords[0] + star.xyCtr[0]) / BIN - 0.25
                 fiber_y = (stampFrameCoords[1] + star.xyCtr[1]) / BIN - 0.25
@@ -1305,6 +1306,14 @@ class GuiderImageAnalysis(object):
                     fiber.fwhm = fwhm / BIN
                     self.cmd.inform('text="gprobe {} has FWHM {:.2f} arcsec"'.format(gprobe.id,
                                                                                      fwhm / BIN))
+
+        if n_tritium == 0:
+            if force is False:
+                msg = 'failed to find a LED source and force=False'
+                self.cmd.error('text="{}"'.format(msg))
+                raise GuiderExceptions.FlatError(msg)
+            else:
+                self.cmd.warn('text="could not find any LED, but force=True so will carry on."')
 
         # Create the processed flat image.
         # NOTE: jkp: using float32 to keep the fits header happier.
@@ -1453,7 +1462,7 @@ class GuiderImageAnalysis(object):
 
         return hdulist, gprobes
 
-    def analyzeFlat(self, flatFileName, gprobes, cmd=None, setPoint=None):
+    def analyzeFlat(self, flatFileName, gprobes, cmd=None, setPoint=None, force=False):
         """
         Return (flat,mask,fibers): with the processed flat, mask to apply
         to the image and flat to mask everything but the fibers, and a list
@@ -1496,7 +1505,7 @@ class GuiderImageAnalysis(object):
                                                                   gprobes, hdr)
             else:
                 hdulist, gprobes = self._find_fibers_in_flat(image, flatFileName,
-                                                             gprobes, hdr)
+                                                             gprobes, hdr, force=force)
 
         elif self.camera == 'ecamera':
             hdulist = self._process_ecam_flat(image, flatFileName, hdr)